Update push-pull-fetch.md
A comma should not be added when there is only one subject for two verbs. For example, in the first edited sentence, the subject is "git push", and its two verbs are "takes" and "checks".  If there were two subjects for the two verbs, you would add the comma.  The same rule applies for the other two edited sentences.